+++ Sources/Sentry/SentryDispatchQueueWrapper.m
@@ -1,0 +1,15 @@
+#import "SentryDispatchQueueWrapper.h"
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@implementation SentryDispatchQueueWrapper
+
+- (void)dispatchAsyncWithBlock:(void (^)(void))block
+{
+    dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0ul), block);
+}
+
+@end
+
+NS_ASSUME_NONNULL_END